+++ src/bench_verify.c
@@ -16,7 +16,7 @@ typedef struct {
     unsigned char msg[32];
     unsigned char key[32];
     unsigned char sig[72];
-    int siglen;
+    size_t siglen;
     unsigned char pubkey[33];
     int pubkeylen;
 } benchmark_verify_t;
